A further typedef bug found by tinne

+++ source/AirwindowsUI.cpp
@@ -36,7 +36,7 @@ void AirwindowsMeter::paint(juce::Graphics &g)
     g.fillRect(0.0,  200.0f*dy, (float)getWidth(),1.0); // border with slew meter
     g.fillRect(0.0,  400.0f*dy, (float)getWidth(),1.0); // border with zero cross meter
 
-    for (u_long count = 0; count < dataPoints; ++count) //count through all the points in the array
+    for (int count = 0; count < dataPoints; ++count) //count through all the points in the array
     {
         g.setColour(juce::Colours::grey);
         g.fillRect((float)dataPosition*dx, 0.0f, 1.0f, (float)getHeight()); //the moving line
